requirements.txt
asgiref==3.5.2
Django==4.1.1
gunicorn==20.1.0
Pillow==9.2.0
sqlparse==0.4.2
tzdata==2022.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
channels==3.0.5